zstd: fold the short-sequence copy test into seqdec_test.go, rewrap a comment
Review feedback: keep sequence decoder tests in seqdec_test.go rather than
adding a file per feature, and restore the line wrap of the
useSafeDecodeSync comment.
