+++ zstd/seqdec_asm.go
@@ -85,8 +85,8 @@ const errorOverread = 6
 // memory copies that read and write in 16-byte blocks, overrunning the end
 // of a literal run or match by up to 15 bytes; an empty literal run still
 // copies one block, reading and writing the full 16 bytes of slack past
-// its position. Otherwise it must use the bounds-exact ("safe") copies. This mirrors the analogous, always-dynamic
-// selection in executeSimple below.
+// its position. Otherwise it must use the bounds-exact ("safe") copies.
+// This mirrors the analogous, always-dynamic selection in executeSimple below.
 //
 // The unsafe copies were disabled in #644 (2022) as a mitigation for a
 // crash, but that crash's root cause — an unguarded bitReader overread that
